+++ tests/Mobizon.Net.Tests/Services/AlphanameServiceTests.cs
@@ -26,14 +26,13 @@ namespace Mobizon.Net.Tests.Services
 
             var result = await CreateService(mockHttp).ListAsync();
 
-            Assert.Equal(MobizonResponseCode.Success, result.Code);
-            Assert.Equal(1, result.Data.TotalItemCount);
-            Assert.Single(result.Data.Items);
-            Assert.Equal(7719L, result.Data.Items[0].Id);
-            Assert.Equal(58356L, result.Data.Items[0].AlphanameId);
-            Assert.Equal(58356L, result.Data.Items[0].Alphaname!.Id);
-            Assert.Equal("Profit", result.Data.Items[0].Alphaname!.Name);
-            Assert.Equal(1, result.Data.Items[0].GlobalStatus);
+            Assert.Equal(1, result.TotalItemCount);
+            Assert.Single(result.Items);
+            Assert.Equal(7719L, result.Items[0].Id);
+            Assert.Equal(58356L, result.Items[0].AlphanameId);
+            Assert.Equal(58356L, result.Items[0].Alphaname!.Id);
+            Assert.Equal("Profit", result.Items[0].Alphaname!.Name);
+            Assert.Equal(1, result.Items[0].GlobalStatus);
         }
     }
 }